community/requirements.txt
alabaster==0.7.12
amqp==1.4.9
anyjson==0.3.3
Babel==2.6.0
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.37.1
django-celery==3.2.2
django-environ==0.4.5
django-redis==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
docutils==0.14
idna==2.7
imagesize==1.1.0
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.0
oauthlib==2.1.0
packaging==18.0
Pillow==5.2.0
Pygments==2.2.0
pyparsing==2.2.2
python3-openid==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
urllib3==1.23
